% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPC.R
\name{OPC}
\alias{OPC}
\title{The online principal component method refers to the IPC method with the best performance among the IPC, the PPC and the SAPC methods.}
\usage{
OPC(data, m, eta)
}
\arguments{
\item{data}{is a highly correlated online data set}

\item{m}{is the number of principal component}

\item{eta}{is the proportion of online data to total data}
}
\value{
Ao,Do
}
\description{
The online principal component method refers to the IPC method with the best performance among the IPC, the PPC and the SAPC methods.
}
\examples{
OPC(data=PSA,m=3,eta=0.8) 
}
