% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.R
\name{PC}
\alias{PC}
\title{The traditional principal component method. This method can estimate the eigen space of the data set.}
\usage{
PC(data, m = m)
}
\arguments{
\item{data}{is a set of highly correlated variables}

\item{m}{is the number of principal component}
}
\value{
Ahat, Dhat
}
\description{
The traditional principal component method. This method can estimate the eigen space of the data set.
}
\examples{
PC(data=PSA,m=3)
}
