\name{PSA}
\alias{PSA}
\docType{data}
\title{
Prostate Specific Antigen
}
\description{
The prostate specific antigen (PSA) data set.
}
\usage{data("PSA")}
\format{
  
  \describe{
    \item{\code{lcavol}}{a numeric vector}
    \item{\code{lweight}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{lbph}}{a numeric vector}
    \item{\code{svi}}{a numeric vector}
    \item{\code{lcp}}{a numeric vector}
    \item{\code{gleason}}{a numeric vector}
    \item{\code{pgg45}}{a numeric vector}
    \item{\code{lpsa}}{a numeric vector}
  }
}
\details{
The data set comes from the prostate specific antigen (PSA)  data of 96 patients collected by Stanford University Medical Center. These patients all underwent radical prostatectomy.
}
\source{
The Stanford University Medical Center.
}
\references{
NA
}
\examples{
data(PSA)
## maybe str(PSA) ; plot(PSA) ...
}
\keyword{datasets}
