% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAPC.R
\name{SAPC}
\alias{SAPC}
\title{The stochastic approximation principal components can handle data sets that are updated in real time and streaming data.}
\usage{
SAPC(data, m, eta)
}
\arguments{
\item{data}{is a highly correlated online data set}

\item{m}{is the number of principal component}

\item{eta}{is the proportion of online data to total data}
}
\value{
Asa,Dsa
}
\description{
The stochastic approximation principal components can handle data sets that are updated in real time and streaming data.
}
\examples{
SAPC(data=PSA,m=3,eta=0.8) 
}
