% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{SPC}
\alias{SPC}
\title{The sparse principal component can obtain sparse solutions of the eigenmatrix to better explain the relationship between principal components and original variables.}
\usage{
SPC(data, m, gamma)
}
\arguments{
\item{data}{is a set of highly correlated variables}

\item{m}{is the number of principal component}

\item{gamma}{is a sparse parameter}
}
\value{
As,Ds
}
\description{
The sparse principal component can obtain sparse solutions of the eigenmatrix to better explain the relationship between principal components and original variables.
}
\examples{
SPC(data=PSA,m=3,gamma=0.03)
}
