% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPC.R
\name{IPC}
\alias{IPC}
\title{The incremental principal components can handle data sets that are updated in real time and streaming data.}
\usage{
IPC(data, m, eta)
}
\arguments{
\item{data}{is a highly correlated online data set}

\item{m}{is the number of principal component}

\item{eta}{is the proportion of online data to total data}
}
\value{
Ai,Di
}
\description{
The incremental principal components can handle data sets that are updated in real time and streaming data.
}
\examples{
IPC(data=PSA,m=3,eta=0.8) 
}
