% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPC.R
\name{OPC}
\alias{OPC}
\title{The online principal components can handle data sets that are updated in real time and streaming data.}
\usage{
OPC(data, m, eta)
}
\arguments{
\item{data}{is a highly correlated online data set}

\item{m}{is the number of principal component}

\item{eta}{is the proportion of online data to total data}
}
\value{
Ao,Do
}
\description{
The online principal components can handle data sets that are updated in real time and streaming data.
}
\examples{
OPC(data=PSA,m=3,eta=0.8) 
}
