% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOPC.R
\name{SOPC}
\alias{SOPC}
\title{The sparse online principal component can not only process the real-time updated data set and stream data set, but also obtain the sparse solution of the updated data set.}
\usage{
SOPC(data, m, gamma, eta)
}
\arguments{
\item{data}{is a highly correlated online data set}

\item{m}{is the number of principal component}

\item{gamma}{is a sparse parameter}

\item{eta}{is the proportion of online data to total data}
}
\value{
Aso,Dso
}
\description{
The sparse online principal component can not only process the real-time updated data set and stream data set, but also obtain the sparse solution of the updated data set.
}
\examples{
SOPC(data=PSA,m=3,gamma=0.03,eta=0.6)
}
